+++ components/ble_profiles/esp/ble_ota/CHANGELOG.md
@@ -1,6 +1,16 @@
+# ChangeLog
+
 ## v0.1.0 - 2023-01-04
 
 ### Feature:
 
 * BLE-OTA:
   * The initial version.
+
+## v0.1.1 - 2023-01-17
+
+### Enhancements:
+* BLE-OTA:
+  * Deleted subscribe command support, no longer needed for protocomm.
+
+  * Added finish command support to indicated end of OTA.